MB-44925 Don't set 'hostname' to the alternate hostname
Prior to this change, we would always populate an empty 'hostname' field
with the hostname which was used to bootstrap the REST client.

This meant that when bootstrapping with an alternate address, alternate
addressing mode would not be enabled causing REST requests to be
dispatched to the correct hostname but the wrong ports.

This change now allows users to setup/test using alternate addresses
with a single node cluster.

Change-Id: I22142a20733904614bc99fac388c8bc8938ed7d2
Reviewed-on: http://review.couchbase.org/c/tools-common/+/148421
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>