MB-45007 Where possible, prioritize the bootstrap node
Prior to this change, requests would be sent to any node which is
running the required service. We will now prioritize sending requests to
the node which we used to bootstrap, note that this only applies when
that node is running the required service.

NOTE: When using a connection string with more than a single host, the
bootstrap node is the one which we successfully bootstrapped against.
This means the boostrap node isn't necessarily the first node in the
connection string supplied to 'NewClient'.

Change-Id: Ic0f7015ebf889fd82f3d363f675b86413a88d5c9
Reviewed-on: http://review.couchbase.org/c/tools-common/+/148686
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Carlos Gonzalez <carlos.gonzalez@couchbase.com>