MB-48295 Fetch the maximum number of vBuckets on the target cluster
We will now fetch the maximum number of vBuckets on the target cluster
after successfully bootstrapping.

Renamed "SkipBootstrap" to "ThisNodeOnly" since the old name was
misleading i.e. we'd still bootstrap, we'd just ignore other nodes.

The "SkipBootstrap" functionality was completely untested, refactored
it slightly to use the existing bootstrap functionality (which is
already tested).

Added targetted unit tests for the "ThisNodeOnly" functionality which
ensures that we ignore other nodes when bootstrapping.

Change-Id: If94ee0d5fc7284371cacf06361de73aa14437777
Reviewed-on: http://review.couchbase.org/c/tools-common/+/160726
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>