MB-48424 Avoid <nil> pointer dereference during request retry
Prior to this change, it was possible for the REST client to panic when
attempting to validate the cluster config from the target host.

This would happen for any requests which failed (with a retryable status
code) prior to fetching the cluster info (which takes place after
bootstrapping).

We will now skip validating the host if we haven't fetched the cluster
info yet.

Change-Id: I7089666e228adb0179e8d8242d79e8815c799767
Reviewed-on: http://review.couchbase.org/c/tools-common/+/161378
Reviewed-by: Patrick Varley <patrick@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>