MB-49039 Add a 'ListParts' function to the 'Client' interface
This function will be used by 'cbbackupmgr' as part of it's resume
validation to ensure that a backup is still resumable.

NOTE: This required refactoring how multipart uploads are handled by the
GCP client so that it's able to keep track of parts via a key and
generated upload id.

Change-Id: Iaed75cf5ebca6628751388fa5ec1fbcb7c4f6a8c
Reviewed-on: https://review.couchbase.org/c/tools-common/+/168256
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Patrick Varley <patrick@couchbase.com>