MB-50734 Correctly allow a zero value 'End' in 'Valid'
Users should be able to supply a zero value 'End' in their 'ByteRange',
this will be translated to the size of the object by all cloud
providers.

NOTE: Fixed cases where we were incorrectly using the 'End' value from
the byte range (with regards to it being a zero value).

Change-Id: I685236102374c72e6a69d83dcbbd9fdb15ea5f1c
Reviewed-on: https://review.couchbase.org/c/tools-common/+/169850
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>