MB-50676 Ensure 'PutObject' requests are treated as idempotent
Prior to this change our 'PutObject' requests were being treated as not
being idempotent and therefore not being retried.

This was because we weren't supplying an object generation, or stating
that we wanted to explicitly create the object.

We now indicate that we should 'RetryAlways' which signals to the SDK
that we want to treat this operation as being idempotent.

NOTE: The normal retry behavior should still be triggered, meaning we
shouldn't be retrying any failures, just transient ones.

Change-Id: If84fd82b91c7257d7b15fd085164277b26443154
Reviewed-on: https://review.couchbase.org/c/tools-common/+/170593
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Maksimiljans Januska <maks.januska@couchbase.com>