MB-50951 Dispatch failed REST requests to another node where possible
Prior to this change it was possible (probable) that a failed REST
request would be re-dispatched to the same node.

This could lead to cases where we'd exhaust our retries against a node
which was actively being removed from the cluster.

We will now shift our traffic across to another node where possible.

Change-Id: I13aa43e3f2788acbcfdadc603ba939ab09595828
Reviewed-on: https://review.couchbase.org/c/tools-common/+/170703
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>