MB-51864 Make `TotalMemory` respect cgroup limits
On linux we want to respect the cgroup memory limit of our process in
order to work better with containers. In order to do this we need to
parse three files:
1. /proc/self/cgroup to give us a directory
2. /proc/self/mountinfo to tell us where that directory is mounted
3. <mount point>/memory.max or memory.limit_in_bytes to tell us the
memory limit

If we fail to parse any of these files, or a memory limit isn't
specified, we fallback to using /proc/meminfo.

Change-Id: I998f1f032f48038b0dd9e188332090df587c3f6a
Reviewed-on: https://review.couchbase.org/c/tools-common/+/173965
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>