MB-52531 Fix Azure client AppendToObject() not working correctly
Prior to this fix AppendToObject() would fail when it's called on
an existing blob when the client was created with static credentials.
Specifically, the blob PUT request would fail with 401 status code
and "Server failed to authenticate the request. Please refer to the
information in the www-authenticate header." error message.

We suspect that this could be due to a bug in the Azure blob storage
SDK. This fix avoids using the potentially broken SDK function but
is not perfect as it requires listing all of the commited blocks of
the blob to which we want to append to so we might need to revisit
this/think of a better solution.

Change-Id: If49c8082024758c9ca7670e938bf149e0c8052f1
Reviewed-on: https://review.couchbase.org/c/tools-common/+/176196
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>