MB-52401 Increase ResponseHeaderTimeout and retry if reached it
Prior to this change REST requests to the CBAS could fail due to
a fairly low default ResponseHeaderTimeout value of our REST HTTP
client. Now this tiemout has been increased to 30 seconds from 10
to alleviate the issue.

Additionally, previously we would not retry idempotent requests
that had failed due to this timeout but would, for example, retry
on failures due to the TLS handshake timeout. Now we retry a
request if we get a standard Go library error that corresponds to
this timeout.

Change-Id: I830c9950ac7b08a8cf50de95fed477dcbf51a038
Reviewed-on: https://review.couchbase.org/c/tools-common/+/176286
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>