MB-52686 Fix cgroup memory limit detection
When MacOS is the host for the container (using docker desktop) it
appears to write a single entry in /proc/self/mountinfo, rather than the
several we expect. For example on docker under Ubuntu 20.04 we get:

1673 1665 0:38 /docker/51722a12fe5ccb03046f01753c92fc06172a5343c6267ec286731ea9e559c476 /sys/fs/cgroup/memory ro,nosuid,nodev,noexec,relatime master:21 - cgroup cgroup rw,memory
1674 1665 0:39 /docker/51722a12fe5ccb03046f01753c92fc06172a5343c6267ec286731ea9e559c476 /sys/fs/cgroup/cpu,cpuacct ro,nosuid,nodev,noexec,relatime master:22 - cgroup cgroup rw,cpu,cpuacct
<more omitted>

but under MacOS we see just:

491 490 0:30 / /sys/fs/cgroup ro,nosuid,nodev,noexec,relatime - cgroup2 cgroup rw

For that reason in this commit we tweak the mountinfo parsing to simply
pick the mount point if there is only one for the given filesystem. If
there are multiple then we do as before and filter by options.

Change-Id: I9b2c8bc7851ac1cf02f052d3d2cde393bc782f8c
Reviewed-on: https://review.couchbase.org/c/tools-common/+/176646
Reviewed-by: Maksimiljans Januska <maks.januska@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>