MB-52686 Respect CPU limits in NumCPU
Use existing cgroup parsing code to work out what the number of CPUs
should be when any cgroup limits have been taken into account. We
fallback to GOMAXPROCS and if GOMAXPROCS has been specified in the
environment we will always use that, even if a cgroup limit has been
specified.

Change-Id: I2537c8d295af7c32703e5c9031d8abae25aa722b
Reviewed-on: https://review.couchbase.org/c/tools-common/+/176654
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>