MB-52686 Fallback to cpu cgroup mount point if cpu,acct is not found
Docker on Linux and Mac appears to group cpu and cpuacct together, but
this is not the case on Docker for Windows using WSLv2.

When looking at the options on a mount point we are now more clever - we
split them into individual options and check if they match the given
system, rather than just checking if the string of all the options
contains the system we are looking for. This is needed so that looking
for "cpu" doesn't match "cpuset" or "cpuacct".

Change-Id: I9dcf7fd7bebf271acf207ed5223c280ab3566891
Reviewed-on: https://review.couchbase.org/c/tools-common/+/176967
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>