MB-52336 Create library for easy up/download to objstore
The library has been designed to allow the services to be able to use it
easily. This means it handles creating a client if you don't provide
one, and has helper functions for up/downloading files and directories.

We already had a lot of this functionality, but not in one easy to use
place.

Change-Id: I8fa227b913aa3bb148aea4929d49ad994fb0c4d1
Reviewed-on: https://review.couchbase.org/c/tools-common/+/176818
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>