MB-53064 Fix deadlock in hofp when queuing
If an error occurs whilst we're still queuing then previously a deadlock
could occur. This is because to queue we simply try and send on a
channel. When a worker goroutine gets an error it exits and causes all
the other workers to exit, meaning there is nothing to drain the
channel. If the buffer of the channel is full then there is a deadlock.

Change-Id: Ied54e3c661d6ea57a05f3f64839327fdaa3a91ee
Reviewed-on: https://review.couchbase.org/c/tools-common/+/177889
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>
Reviewed-by: Maksimiljans Januska <maks.januska@couchbase.com>