MB-52612 Fix potentially flaky ratelimit tests
In CI we sometimes see tests run slow. Therefore checking that a
read/write/s3 sync takes about as long as the rate limit would suggest
can be flaky. Instead we just make sure that it takes at least that
long.

Unfortunately it's not quite as easy as this. Doing test setup takes a
little bit of time which means we've already waited a small proportion
of the rate limit time before we register the start time, so we end up
taking slightly less time to run the method than we think. We allow some
leeway here.

Ultimately we're trading off one source of flakiness here for another.
Given we've seen the flakiness this commit addresses it seems a
reasonable change.

Change-Id: I5cebd8d5dbf1020aa1bc2c0b1afe7633aa3ef04c
Reviewed-on: https://review.couchbase.org/c/tools-common/+/178700
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>