MB-53826 Add freelist data structure
A freelist is a structure which is commonly used as a pool of
pre-allocated objects. Go has something similar - sync.Pool - but it
does not allow one to cap the number of objects in it. This is a useful
feature when memory allocation needs to be limited.

The implementation in this commit is a simple wrapper over a Go channel.
A more custom implementation was also tried and benchmarked but it
provided much worse performance when contention was low, and only
marginally better performance when it was medium to high. For our use
cases it is likely that contention is going to be low.

Change-Id: I05b64ab85ff6c8ee4ed94f80664bc88d58c68490
Reviewed-on: https://review.couchbase.org/c/tools-common/+/180452
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>