MB-53826 Add objutil.CompressObjects funtion
This function can be used to zip up a number of objects (specified in
the same way as Client.IterateObjects) and upload that zip to cloud
storage on the fly.

It is currently largely sequential - a file is downloaded, it's zipped
and periodically the zip is flushed to cloud storage.

Change-Id: I67bfa1f1db47de3e3748c4f12ccffc31bdf3a110
Reviewed-on: https://review.couchbase.org/c/tools-common/+/180453
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>