MB-53958 Fix pre-commit script
The pre-commit script currently exits with an error silently
when the git parameter couchbase.tools.format.behaviour is not set,
because the attempt to assign it to config using git config
results in an error. This change will set config to an
empty string ("") if couchbase.tools.format.behaviour is
unset.

Change-Id: I460f8b5f9e113f147dd508cf597260e4a781d0bf
Reviewed-on: https://review.couchbase.org/c/tools-common/+/182537
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>