MB-54627 Ensure deleteObjects doesn't error when no keys exist
When using the DeleteDirectory method on objaws.Client it should
delete a "directory" in S3. If you do this on a directory with no
objects in it an error is returned because deleteObjects sends an
S3 request even if there are no keys to delete. To fix this we
check if there are any keys, if not we exit early.

Change-Id: I6b3807ad51cf12643293c1962e551325816152f2
Reviewed-on: https://review.couchbase.org/c/tools-common/+/183084
Reviewed-by: Tom Tonner <tom.tonner@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>