MB-55201 Upgrade to v0.5.1 of the Azure SDK and use gomock
We need to upgrade so that we can use objstore in Capella which is on
this newer version.

A lot has changed between v0.4.1 and v0.5.1 of the SDK so this change is
quite intrusive. Most of the changes are moving things around and small
changes to method signatures with a few exceptions:

1. There is no longer a way to see if a client was created using a
   shared key. We need this in order to know whether we need to create a
   SAS URL when copying a part.

2. Some types are no longer exported. This largely seems to be an
   oversight (see https://github.com/Azure/azure-sdk-for-go/issues/19775)
   from moving the types to an internal package and then exporting them
   using type aliases.

3. Generics are everywhere

The first problem is worked around by trying GetSASURL and then handling
the error if there is one. To handle the error we have to do a string
comparison as it is not exported; they have fixed it, but we cannot pick
the change up yet (see
https://github.com/Azure/azure-sdk-for-go/issues/19864).

The second problem only affects us twice. Once for a wrapper around a
string (`BlobPrefix`), so we just use a string. The other time is for
a result type of paging. This is more problematic as we would like to
test this behaviour. This will be fixed in an upcoming verison (see the
linked SDK issue) but until then I have removed the tests around this.

The second are third problem are linked. Because there are generics
everywhere we have had to use extra interfaces with concrete types to be
able to mock. Unfortunately the mock library we were using (testify)
follows type aliases and so tries to import the internal package. To
rectify this we have had to switch to a different mock library - gomock
- which is used in the cloud organisation.

Change-Id: I93aa485a30bac08c47d9c46ea9533c24d7cdd7de
Reviewed-on: https://review.couchbase.org/c/tools-common/+/185465
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>