MB-55967 Fix improper error handling
CompressObjects creates 2 seperate worker threads, one for downloading
data, and one for zipping it and uploading it. The downloading thread
writes to a ZipWriter, and the zipping/uploading thread reads from it.

Previously, if the downloading thread hit an error, it wouldn't close
the ZipWriter, meaning the other thread would hang, indefinitely
waiting for something to either be written to the ZipWriter. This
commit fixes that, and ensures that both threads are closed properly
in the event of the download erroring out.

Change-Id: I21157f104d6ce9548adaf51046ab495fedd64406
Reviewed-on: https://review.couchbase.org/c/tools-common/+/188374
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>