MB-56478 Provide progress reports during `CompressUpload`
If the user provides a callback then `CompressUpload` will call it when
a download of an object has complete with the progress through the total
amount to be downloaded.

This is done by first calculating the total size of all the objects with
the path prefix. We then keep a running total of the number of bytes
that have been downloaded which is updated every time a download
completes.

This approach was chosen due to its simplicity but it does mean the
rate of progress reports depends on the size of the objects in cloud
storage. If a user has extremely big objects in cloud storage
(or very few) this may not be ideal, but for our purposes it is
sufficient.

Change-Id: Icd38911d0a333c2f7f3867d64444837bd8084cce
Reviewed-on: https://review.couchbase.org/c/tools-common/+/189716
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Safian Ali <safian.ali@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>
Reviewed-by: Maksimiljans Januska <maks.januska@couchbase.com>