MB-56818 Create a new package httptools
This package copies most of its functionality from cbrest but aims
to make it more general and applicable to more use cases.

The purpose of this is to have a tool to make requests to any endpoint
with robust retries built in so we do not need to litter the codebase
with our own retryers.

Change-Id: I130fadebf918676538aac79e66eb7c623f10776c
Reviewed-on: https://review.couchbase.org/c/tools-common/+/190765
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Reviewed-by: Maksimiljans Januska <maks.januska@couchbase.com>