MB-56728 On S3 uploads use trailing checksums to avoid double reads
By default the S3 SDK reads the data you are uploading once to checksum
it and once more to actually send it. By specifying a checksum algorithm
we can change this behaviour so that it calculates it as data is
uploaded.

SHA-256 was chosen as this is the algorithm the SDK uses when the
parameter is not set. In terms of integrity checking
this means the change does not alter behaviour.

Change-Id: Id01446e11c6824d0d6aed0511f25d04b26df75df
Reviewed-on: https://review.couchbase.org/c/tools-common/+/190905
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Maksimiljans Januska <maks.januska@couchbase.com>