MB-56890 Clean up cbrest client test resources
If we were to run all tests for the cbrest package some would fail or
hang because the resource closing and clean up was omitted in some tests
which caused issues for the later tests.

This change makes sure we call the Close method for all the resources
that were created successfuly and need to be cleaned up after the test
is done.

Change-Id: Ia82129e3f6a472e135273ef7188273a01fad4470
Reviewed-on: https://review.couchbase.org/c/tools-common/+/190963
Reviewed-by: Matt Hall <matt.hall@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>