MB-57297 Fix hang in `objutil.CompressUpload`
The hang would happen when the upload goroutine returned an error. In
normal operation this goroutine would read out of the `ZipWriter`. With
this stopped the writer would fill up its buffer, meaning no more bytes
could be written into it. In the download goroutine we were attempting
to read out of S3 and into the `ZipWriter` using an `io.ReadAll`. This
call is blocking (i.e. does not take a `Context`) so the writer being
full meant this goroutine would never terminate.

The fix is to close the reader the upload goroutine is reading out of.
This causes an error in `io.ReadAll` which allows the download goroutine
to terminate.

This was originally found when we hit the limit on the number of parts
in an upload on AWS (10,000).

Change-Id: Iad2d7ce31ed7de90af41cdd41433afec7dbdd6b3
Reviewed-on: https://review.couchbase.org/c/tools-common/+/192304
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>