MB-62154 Fix `TestClient.IterateObjects` for nested keys
Previously we fixed a couple of bugs in `TestClient.IterateObjects`
where duplicate entries were returned and partial paths could be
returned. Unfortunately this regressed behaviour when you have a nested
path (e.g. "foo/bar/baz", not "foo") - we would return directory stubs
for them.

NOTE: The version that introduced the regression was never pushed, so we
do not need to bump any version numbers.

Change-Id: I598a1078b13978c91c42ddc2ad79baf93854fa23
Reviewed-on: https://review.couchbase.org/c/tools-common/+/212299
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Safian Ali <safian.ali@couchbase.com>