MB-62846 Add script to generate metrics code & metadata
Previously we had code in the backup service which would generate
boilerplate Prometheus code and the metrics_metadata.json file. This was
very specific to the backup service however - e.g. the generation
script worked by importing a list of metrics defined in Go code.

This commit introduces a more general metrics metadata & code generation
tool. It uses a JSON description of metrics which is a small extension
to the metadata format. Namely we introduce a `histogram_buckets` field
which is needed to configure a histogram metric. The tool will then
output code and metadata JSON (stripped of the extra field).

Change-Id: Ic241b68e72f3321cb9239d522677bc645ff3992b
Reviewed-on: https://review.couchbase.org/c/tools-common/+/213109
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Safian Ali <safian.ali@couchbase.com>