MB-65094 Use CRC32 for checksum of S3 PUTs
The previous SDK always sent an MD5 checksum on PUTs. The newer SDK does
not send any by default, which causes problems for object lock-enabled
buckets where a checksum is required.

Testing
-------
Ran a backup to an S3 bucket with object lock turned on, with and
without this change. It failed without it, completed with it.

Change-Id: Ifbefa3404cdf30af9735fc43aa4e1748c91a4117
Reviewed-on: https://review.couchbase.org/c/tools-common/+/222713
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>