MB-65237 Move ReplaceLocalhost from cbbs repo to TC
In the backup service we use this to fixup a cbauth idiosyncrasy. Using
`cluster_run` it is common to get a node with a loopback node name. If
you then try to get credentials for the hostname `localhost` it will
always return with credentials for the node with a loopback node name,
even if you are on another node.

This was fixed in cbauth but ultimately rolled back by the ns_server
team as it broke a test script of another service.

Change-Id: I2b64abfdda5c507ceb2ec7c2d30553aac7f0410a
Reviewed-on: https://review.couchbase.org/c/tools-common/+/225871
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>