MB-66135 Don't hang if certs change
When a request fails, the main thread will block waiting for `pollCC`
to update the cluster config. If pollCC is never able to update the
cluster config e.g. because the cluster's certs were rotated after the
client was started, the main thread will block indefinitely.

To prevent this, we cancel the context if we fail to update the cluster
config after too many retries.

Change-Id: If9416eafc94aa9895b47c637a638756891301fee
Reviewed-on: https://review.couchbase.org/c/tools-common/+/225911
Reviewed-by: James Lee <james.lee@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>