MB-66135 Add timeout to WaitUntilUpdated
When a request fails, the main thread will block waiting for `pollCC`
to update the cluster config. If pollCC is never able to update the
cluster config e.g. because the cluster's certs were rotated after the
client was started, the main thread will block indefinitely.

To prevent this, we add a 10s timeout to WaitUntilUpdated.

Change-Id: If9416eafc94aa9895b47c637a638756891301fee