MB-67927 Add If-Match support to objstore
Each cloud provider now supports a CAS operation using the If-Match
header. Unfortunately, AWS and Azure use the ETag and GCP uses the
"generation" (i.e. the version ID).

To get around this we update `ObjectAttrs` to have a `CAS` field, which
is set to the correct value to compare for each cloud provider.

Testing
-------
On each cloud provider:

1. Create an object
2. Get the "CAS" value
3. Write with "If-Match" set to the CAS value from 2 (succeeds)
4. Write again with "If-Match" set to the CAS value from 3 (fails)

Change-Id: I6a6a636cebfd612a256d0df9e4768fe2e801105c
Reviewed-on: https://review.couchbase.org/c/tools-common/+/231787
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Lee <james.lee@couchbase.com>