MB-69002 Add `MultiServer` to couchbase/rest
This can be used by Couchbase services to listen on multiple addresses
serving multiple protocols using one type. This is useful as Couchbase
services have to listen on ipv4 and/or ipv6 and accept http and/or https
depending on server settings, which can change. This package contains
this functionality in `MultiServer`.

Testing
-------
Used in the backup service and turned on ipv6 only mode.

Change-Id: I6a6a636c5348e31ce88f03d1d955f6e59735c8b6
Reviewed-on: https://review.couchbase.org/c/tools-common/+/235007
Reviewed-by: Safian Ali <safian.ali@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>