MB-66940 Add delete markers to GCP IterateObjects
GCP does not provide a mechanism for determining the current (live)
version of an object when iterating object versions. However, we can
use the objects' `Deleted` time to infer whether the latest version
if current. To keep our API consistent we will simulate delete markers
similar to the ones used by AWS.

Change-Id: I3bebfd6225b97c5205c7f6e859cf035c24be06b6
Reviewed-on: https://review.couchbase.org/c/tools-common/+/235090
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>