MB-69551 Add a function to check whether object versioning is enabled
Add a `GetBucketVersioningStatus` function to check whether object
versioning is enabled for a specified bucket. We currently support this
only for S3 and GCP. For Azure we always return an
`UnimplementedOperationError`.

NOTE: It is possible to implement this for Azure as well, however we do
not currently need it so it is not worth it just for consistency. On
Azure versioning is enabled at the storage account level so we cannot
do this through azblob. We would need to create (and authenticate) a
separate client to access the storage account.

Change-Id: I22ecd68f6a9558113d492b943e6f26fad575113a
Reviewed-on: https://review.couchbase.org/c/tools-common/+/237142
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Matt Hall <matt.hall@couchbase.com>